Cambio en tipo de variables de INT a FLOAT

+++ src/main/java/com/aguatzi/sistemaAguatzi/entidades/CierreCajaRepartidor.java
@@ -31,19 +31,19 @@ public class CierreCajaRepartidor implements Serializable {
     private int pagadosTransferencia;
 
     @Column(name = "dineroReportado", nullable = false)
-    private int dineroReportado;
+    private float dineroReportado;
 
     @Column(name = "gastosExtra", nullable = false)
-    private int gastosExtra;
+    private float gastosExtra;
 
     @Column(name = "especificaciones",length = 150, nullable = false)
     private String especificaciones;
 
     @Column(name = "dineroVentaTotal", nullable = false)
-    private int dineroVentaTotal;
+    private float dineroVentaTotal;
 
     @Column(name = "dineroFaltante", nullable = false)
-    private int dineroFaltante;
+    private float dineroFaltante;
 
     @ManyToOne
     @JoinColumn(name = "idEmpleado", nullable = false)
@@ -97,7 +97,7 @@ public class CierreCajaRepartidor implements Serializable {
         this.pagadosTransferencia = pagadosTransferencia;
     }
 
-    public int getDineroReportado() {
+    public float getDineroReportado() {
         return dineroReportado;
     }
 
@@ -105,7 +105,7 @@ public class CierreCajaRepartidor implements Serializable {
         this.dineroReportado = dineroReportado;
     }
 
-    public int getGastosExtra() {
+    public float getGastosExtra() {
         return gastosExtra;
     }
 
@@ -121,7 +121,7 @@ public class CierreCajaRepartidor implements Serializable {
         this.especificaciones = especificaciones;
     }
 
-    public int getDineroVentaTotal() {
+    public float getDineroVentaTotal() {
         return dineroVentaTotal;
     }
 
@@ -129,7 +129,7 @@ public class CierreCajaRepartidor implements Serializable {
         this.dineroVentaTotal = dineroVentaTotal;
     }
 
-    public int getDineroFaltante() {
+    public float getDineroFaltante() {
         return dineroFaltante;
     }
